shuffle rebase
